+++ packages/core/src/helpers/dom.ts
@@ -17,3 +17,16 @@
 export function getAttributeValues(target: Element, attributeName: string): string[] {
   return target.getAttribute(attributeName)?.trim().split(/\s+/) || [];
 }
+
+/**
+ * Returns a promise that is fulfilled when the document's `readyState` is `complete`.
+ */
+export function domReady() {
+  return new Promise<void>((resolve) => {
+    if (document.readyState === 'loading') {
+      document.addEventListener('DOMContentLoaded', () => resolve());
+    } else {
+      resolve();
+    }
+  });
+}